Make all tests use the same index deletion code
